AutoGenchain - Kali/requirements.txt
langchain==0.0.292
langchain-experimental==0.0.35
pyautogen==0.2.19
chromadb==0.4.18
openai==1.3.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3 